pyproject.toml
[tool.poetry]
name = "pydev-demo"
version = "0.1.0"
description = ""
authors = ["Eddie <eddie@5xcampus.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-extensions = "^3.2.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
